Thought maybe initing the boostrap LAPIC might fix the issue, apparently not

+++ kernel/src/arch/x86_64/cpu/ioapic.c
@@ -60,7 +60,7 @@ void ioapic_redirect_gsi(uint32 lapic_id, uint8 vector, uint32 gsi, uint16 flags
     redirect |= (uint64_t)lapic_id << 56;
 
     uint32 redir_table = (gsi - ioapic->gsi_base) * 2 + 16;
-    write_ioapic(ioapic, redir_table, redirect);
+    write_ioapic(ioapic, redir_table, (uint32)redirect);
     write_ioapic(ioapic, redir_table + 1, redirect >> 32);
 }
